+++ src/HCFS/file_present.h
@@ -24,46 +24,46 @@
 #include "meta_mem_cache.h"
 #include "xattr_ops.h"
 
-int fetch_inode_stat(ino_t this_inode, struct stat *inode_stat,
-		unsigned long *ret_gen, char *ret_pin_status);
+int32_t fetch_inode_stat(ino_t this_inode, struct stat *inode_stat,
+		uint64_t *ret_gen, char *ret_pin_status);
 
-int mknod_update_meta(ino_t self_inode, ino_t parent_inode,
+int32_t mknod_update_meta(ino_t self_inode, ino_t parent_inode,
 			const char *selfname,
-			struct stat *this_stat, unsigned long this_gen,
-			ino_t root_ino, long long *delta_meta_size, char ispin);
+			struct stat *this_stat, uint64_t this_gen,
+			ino_t root_ino, int64_t *delta_meta_size, char ispin);
 
-int mkdir_update_meta(ino_t self_inode, ino_t parent_inode,
+int32_t mkdir_update_meta(ino_t self_inode, ino_t parent_inode,
 			const char *selfname,
-			struct stat *this_stat, unsigned long this_gen,
-			ino_t root_ino, long long *delta_meta_size, char ispin);
+			struct stat *this_stat, uint64_t this_gen,
+			ino_t root_ino, int64_t *delta_meta_size, char ispin);
 
-int unlink_update_meta(fuse_req_t req, ino_t parent_inode,
+int32_t unlink_update_meta(fuse_req_t req, ino_t parent_inode,
 			const DIR_ENTRY *this_entry);
 
-int meta_forget_inode(ino_t self_inode);
+int32_t meta_forget_inode(ino_t self_inode);
 
-int rmdir_update_meta(fuse_req_t req, ino_t parent_inode, ino_t this_inode,
+int32_t rmdir_update_meta(fuse_req_t req, ino_t parent_inode, ino_t this_inode,
 			const char *selfname);
 
-int symlink_update_meta(META_CACHE_ENTRY_STRUCT *parent_meta_cache_entry,
+int32_t symlink_update_meta(META_CACHE_ENTRY_STRUCT *parent_meta_cache_entry,
 	const struct stat *this_stat, const char *link,
-	const unsigned long generation, const char *name,
-	ino_t root_ino, long long *delta_meta_size, char ispin);
+	const uint64_t generation, const char *name,
+	ino_t root_ino, int64_t *delta_meta_size, char ispin);
 
 int fetch_xattr_page(META_CACHE_ENTRY_STRUCT *meta_cache_entry,
-	XATTR_PAGE *xattr_page, long long *xattr_pos, BOOL create_page);
+	XATTR_PAGE *xattr_page, int64_t *xattr_pos, BOOL create_page);
 
-int link_update_meta(ino_t link_inode, const char *newname,
-	struct stat *link_stat, unsigned long *generation,
+int32_t link_update_meta(ino_t link_inode, const char *newname,
+	struct stat *link_stat, uint64_t *generation,
 	META_CACHE_ENTRY_STRUCT *parent_meta_cache_entry);
 
-int increase_pinned_size(long long *reserved_pinned_size,
-		long long file_size);
+int32_t increase_pinned_size(int64_t *reserved_pinned_size,
+		int64_t file_size);
 
-int decrease_pinned_size(long long *reserved_release_size,
-		long long file_size);
+int32_t decrease_pinned_size(int64_t *reserved_release_size,
+		int64_t file_size);
 
-int pin_inode(ino_t this_inode, long long *reserved_pinned_size);
-int unpin_inode(ino_t this_inode, long long *reserved_release_size);
+int32_t pin_inode(ino_t this_inode, int64_t *reserved_pinned_size);
+int32_t unpin_inode(ino_t this_inode, int64_t *reserved_release_size);
 
 #endif /* GW20_HCFS_FILE_PRESENT_H_ */
